refactor router config to it's own module, not in
main.rs any longer.
