+++ src/bin/main.rs
@@ -1,53 +1,30 @@
 extern crate httparse;
 extern crate cautious_eureka;
+
 use cautious_eureka::thread_pool::ThreadPool;
 use cautious_eureka::router;
 use self::httparse::Request;
 
+pub mod routes;
+
 use std::io::prelude::*;
 use std::net::TcpListener;
 use std::net::TcpStream;
-// Commented out due to line 21 no longer needing it
-// use std::sync::Arc;
-// use std::sync::Mutex;
 
 fn main() {
     let listener = TcpListener::bind("127.0.0.1:8080").unwrap();
     let pool = ThreadPool::new(4);
-    // Commented out until I can work out why this is not working for 
-    // threading issue, temporary fix on line 24 in just using a new
-    // router per request
-    // 
-    // let server_router = Arc::new(Mutex::new(router_config()));
 
     for stream in listener.incoming() {
         let stream = stream.unwrap();
 
         pool.execute(|| {
-            handle_connection(stream, router_config());
+            handle_connection(stream, routes::config());
         });
     }
 
     println!("Shutting down.");
 }
-
-fn router_config() -> router::Router {
-    let mut routes: Vec<router::Route> = Vec::new();
-
-    let mut headers = [httparse::EMPTY_HEADER; 16];
-    let req = Request::new(&mut headers);
-    let index_func = index(&req);
-    let index = make_route(
-        String::from("/"),
-        move |_| index_func.clone()
-    );
-    routes.push(
-        index
-    );
-
-    return router::Router::new(routes);
-}
-
 
 fn handle_connection(
     mut stream: TcpStream,
@@ -69,13 +46,3 @@ fn handle_connection(
     stream.flush().unwrap();
 }
 
-fn make_route<F>(route: String, func: F) -> router::Route
-    where
-    F: Fn(&Request) -> router::ResponseObject + 'static
-{
-    return router::Route::new(route, func);
-}
-
-fn index(_request: &Request) -> router::ResponseObject {
-    return router::ResponseObject::new(String::from("HTTP/1.1 200 OK\r\n\r\n"), String::from("hello.html"));
-}
